pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["aiflows"]

[project]
name = "aiflows"
version = "0.0.1"
description = 'aiFlows embodies the abstraction of Flows proposed in "Flows: Building Blocks of Reasoning and Collaborating AI" by Josifoski et al., and greatly simplifies the design and implementation of complex (work)Flows involving humans, AI systems, and tools.'  
authors = [
  { name = "ToDo", email = "my_email@domain.com" },  # ToDo: Is this necessary? If so put names.
]
license = {file = "LICENSE"}

dependencies = [
    "colorama==0.4.6",
    "pytest==7.3.1",
    "pytest-cov==4.1.0",
    "hydra-core==1.2.0",
    "hydra-colorlog==1.1.0",
    "pyext==0.7",
    "wrapt-timeout-decorator==1.3.12.2",
    "diskcache==5.6.1",
    "openai==1.0.0",
    "huggingface_hub==0.15.1",
    "jsonlines==3.1.0",
    "jinja2==3.1.2",
    "mock==5.0.2",
    "rich==12.6.0",
    "litellm==1.0.0"
]

# ToDo: check links
[project.urls]
Home = "https://epfl-dlab.github.io/flows/"
Documentation = "https://epfl-dlab.github.io/flows/docs/built_with_sphinx/html/index.html"  
Source = "https://github.com/epfl-dlab/flows"
